+++ examples/collision.rs
@@ -6,6 +6,9 @@ use rusty_engine::prelude::*;
 
 const ROTATION_SPEED: f32 = 3.0;
 
+#[derive(Resource)]
+struct GameState {}
+
 fn main() {
     let mut game = Game::new();
     let msg2 = game.add_text(
@@ -15,7 +18,7 @@ fn main() {
     msg2.font_size = 20.0;
     msg2.translation.y = 340.0;
 
-    let mut race_car = game.add_sprite("Player", SpritePreset::RacingCarGreen);
+    let race_car = game.add_sprite("Player", SpritePreset::RacingCarGreen);
     race_car.translation = Vec2::new(0.0, 0.0);
     race_car.rotation = UP;
     race_car.layer = 100.0;
@@ -28,20 +31,20 @@ fn main() {
                 break 'outer;
             }
             let sprite_preset = sprite_presets_iter.next().unwrap();
-            let mut sprite = game.add_sprite(format!("{:?}", sprite_preset), sprite_preset);
+            let sprite = game.add_sprite(format!("{:?}", sprite_preset), sprite_preset);
             sprite.translation = Vec2::new(x as f32, (-y) as f32);
             sprite.collision = true;
         }
     }
 
-    let mut text = game.add_text("collision text", "");
+    let text = game.add_text("collision text", "");
     text.translation = Vec2::new(0.0, -200.0);
 
     game.add_logic(logic);
-    game.run(());
+    game.run(GameState {});
 }
 
-fn logic(engine: &mut Engine, _: &mut ()) {
+fn logic(engine: &mut Engine, _: &mut GameState) {
     // If a collision event happened last frame, print it out and play a sound
     for collision_event in engine.collision_events.drain(..) {
         let text = engine.texts.get_mut("collision text").unwrap();